# src/main/resources/api.yaml
openapi: 3.0.3
servers:
  - url: http://localhost:8080
info:
  title: Restaurant Review REST API
  version: 1.0.0
  description: Spring Boot Restaurant Review REST API Documentation
  contact:
    name: Elizabeth Drew
paths:
  /api/v1/search:
    get:
      summary: Search all restaurants
      operationId: getSearch
      tags:
        - Search
      description: Returns a list of restaurant results.
      x-spring-paginated: true
      parameters:
        - in: query
          name: query
          required: true
          description: Search query to filter restaurants.
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Restaurant'
        '404':
          description: No Results Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/users:
    post:
      summary: Add a new user
      tags:
        - User
      description: Adds a new user to the database. Valid passwords should contain 8 to 100 characters, at least one uppercase, lowercase, digit and special character.
      operationId: addUser
      security:
        - basicAuth: []
      requestBody:
        description: The user to add.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid Input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Already Exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/users/{userId}:
    get:
      summary: Get a user by ID
      tags:
        - User
      description: Returns a single user by ID
      operationId: getUserById
      security:
        - basicAuth: []
      parameters:
        - name: userId
          in: path
          description: The ID or name of the user to retrieve.
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update a user
      tags:
        - User
      description: Updates a user by ID.
      operationId: updateUserById
      security:
        - basicAuth: []
      parameters:
        - name: userId
          in: path
          description: The ID or name of the user to retrieve.
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        description: The updated user.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid Input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a user by ID
      tags:
        - User
      operationId: deleteUserById
      security:
        - basicAuth: []
      parameters:
        - name: userId
          in: path
          description: The ID or name of the user to retrieve.
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '204':
          description: Deleted
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/reviews:
    get:
      security: []
      summary: Get all reviews
      operationId: getAllReviews
      tags:
        - Reviews
      description: Returns a list of all reviews.
      x-spring-paginated: true
      parameters:
        - name: restaurant_id
          in: query
          description: The restaurant ID to filter reviews by.
          required: false
          schema:
            type: integer
            format: int64
        - name: user_id
          in: query
          description: The user ID to filter reviews by.
          required: false
          schema:
            type: integer
            format: int64
        - name: rating
          in: query
          description: The rating value to filter restaurants by.
          required: false
          schema:
            type: array
            items:
              type: integer
              minimum: 1
              maximum: 5
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        '404':
          description: No Results Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Add a new review
      operationId: addNewReview
      tags:
        - Reviews
      description: Adds a new review to the database. User cannot review a restaurant more than once within a year. Owner of restaurant cannot review their own restaurant.
      security:
        - basicAuth: [ ]
      requestBody:
        description: The review to add.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Invalid Input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Already Exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/reviews/{reviewId}:
    get:
      security: []
      summary: Get a review by ID
      operationId: getReviewById
      tags:
        - Reviews
      description: Returns a single review by ID.
      parameters:
        - name: reviewId
          in: path
          description: The ID of the review to retrieve.
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update a review
      operationId: updateReviewById
      tags:
        - Reviews
      description: Updates a review by ID.
      security:
        - basicAuth: [ ]
      parameters:
        - name: reviewId
          in: path
          description: The ID of the review to update.
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        description: The updated review data.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Invalid Input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a review
      operationId: deleteReviewById
      description: Deletes a review by ID.
      security:
        - basicAuth: [ ]
      tags:
        - Reviews
      parameters:
        - name: reviewId
          in: path
          description: The ID of the review to delete.
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '204':
          description: Deleted
        '403':
          description: Insufficient Permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
        restaurant_id:
          type: integer
          format: int64
        restaurant_name:
          type: string
        restaurant_city:
          type: string
        user_id:
          type: integer
          format: int64
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_deleted:
          type: boolean
          default: false
    ReviewInput:
      type: object
      properties:
        restaurant_id:
          type: integer
          format: int64
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required:
        - restaurant_id
        - rating
        - comment
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The unique identifier of the user.
        name:
          type: string
          description: The name of the user.
        email:
          type: string
          format: email
          description: The email address of the user.
        username:
          type: string
          description: The username of the user.
        password:
          type: string
          description: The password of the user.
        role:
          type: string
          description: The role of the user (admin or reviewer)
          enum: ['ADMIN', 'REVIEWER']
        created_at:
          type: string
          format: date-time
          description: The date and time the user was created.
        is_deleted:
          type: boolean
          default: false
    UserInput:
      type: object
      properties:
        name:
          type: string
          description: The name of the user.
        email:
          type: string
          format: email
          description: The email address of the user.
        username:
          type: string
          description: The username of the user.
        password:
          type: string
          description: The password of the user.
        role:
          type: string
          description: The role of the user (admin or reviewer)
          enum: [ 'ADMIN', 'REVIEWER' ]
          default: 'REVIEWER'
      required:
        - name
        - email
        - password
        - username
        - role
    Restaurant:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The unique identifier of the restaurant.
        name:
          type: string
          description: The name of the restaurant.
        city:
          type: string
          description: The city where the restaurant is located.
        rating:
          type: integer
          minimum: 1
          maximum: 5
          description: The rating of the restaurant.
        price_range:
          type: integer
          minimum: 1
          maximum: 3
          description: The price range of the restaurant.
        latitude:
          type: number
          format: double
          description: The latitude coordinates of the restaurant.
        longitude:
          type: number
          format: double
          description: The longitude coordinates of the restaurant.
        cuisines:
          type: array
          items:
            type: string
            description: List of cuisine names the restaurant offers
        created_at:
          type: string
          format: date-time
          description: The date and time the restaurant was created.
        is_deleted:
          type: boolean
          default: false
        is_favourite:
          type: boolean
          default: false
        total_favourites:
          type: integer
          format: int64
        owner_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
...
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
      required:
        - code
        - message
...